created trust registry configuration module; removed provider interface for configuration; one class per file
