--- src/common/enums/providers.ts
@@ -1,1 +1,0 @@
-export const CREDENTIAL_CONFIG_YML_ADAPTER = 'CREDENTIAL_CONFIG_YML_ADAPTER';
