+++ src/domain/agent/agent/agent.service.ts
@@ -8,7 +8,6 @@ import {
   ValidationException,
 } from '@common/exceptions';
 import { SsiException } from '@common/exceptions/ssi.exception';
-import { CredentialMetadata } from '@services/platform/trust-registry-adapter/credentials/credential.provider.interface';
 import {
   Agent,
   CreateAgentInput,
@@ -34,7 +33,6 @@ import { Cache } from 'cache-manager';
 import { WINSTON_MODULE_NEST_PROVIDER } from 'nest-winston';
 import { firstValueFrom } from 'rxjs';
 import { FindOneOptions, Repository } from 'typeorm';
-import { IClaim } from '@services/platform/trust-registry-adapter/claim/claim.entity';
 import { CredentialService } from '../credential/credential.service';
 import { RestEndpoint } from '@common/enums/rest.endpoint';
 import { WalletManagerCommand } from '@common/enums/wallet.manager.command';
@@ -42,6 +40,8 @@ import { BeginCredentialRequestOutput } from '../credential/dto/credential.reque
 import { BeginCredentialOfferOutput } from '../credential/dto/credential.offer.dto.begin.output';
 import { CredentialMetadataOutput } from '../credential/dto/credential.dto.metadata';
 import jwt_decode from 'jwt-decode';
+import { IClaim } from '@services/platform/trust-registry-adapter/claim/claim.interface';
+import { CredentialMetadata } from '@services/platform/trust-registry-adapter/configuration/credential.metadata';
 
 @Injectable()
 export class AgentService {
