+++ src/services/platform/trust-registry-adapter/claim/claim.community.member.ts
@@ -1,0 +1,17 @@
+import { SystemClaims } from './claim.enum';
+import { IClaim } from './claim.interface';
+
+// The initial idea was to perform validations in the resolvers and derive the claim
+export class CommunityMemberClaim implements IClaim {
+  private communityID: string;
+
+  constructor({ communityID }: { communityID: string }) {
+    this.communityID = communityID;
+  }
+
+  asClaimObject(): Record<string, any> {
+    return {
+      [`${SystemClaims.CommunityMember}_communityID`]: this.communityID,
+    };
+  }
+}
