+++ src/services/platform/trust-registry-adapter/claim/claim.interface.ts
@@ -1,0 +1,3 @@
+export interface IClaim {
+  asClaimObject(): Record<string, any>;
+}
