+++ src/services/platform/trust-registry-adapter/claim/claim.service.ts
@@ -1,6 +1,6 @@
 import { Inject, Injectable, LoggerService } from '@nestjs/common';
 import { WINSTON_MODULE_NEST_PROVIDER } from 'nest-winston';
-import { IClaim } from './claim.entity';
+import { IClaim } from './claim.interface';
 
 @Injectable()
 export class ClaimService {
