+++ src/services/platform/trust-registry-adapter/configuration/trust.registry.configuation.module.ts
@@ -1,0 +1,8 @@
+import { Module } from '@nestjs/common';
+import { TrustRegistryConfigurationAdapter } from './trust.registry.configuration.adapter';
+
+@Module({
+  providers: [TrustRegistryConfigurationAdapter],
+  exports: [TrustRegistryConfigurationAdapter],
+})
+export class TrustRegistryConfigurationModule {}
