+++ src/services/platform/trust-registry-adapter/configuration/trust.registry.configuration.ts
@@ -1,0 +1,5 @@
+import { CredentialMetadata } from './credential.metadata';
+
+export type TrustRegistryConfiguration = {
+  credentials: CredentialMetadata[];
+};
