--- src/services/platform/trust-registry-adapter/credential-config-adapter/credential.config.adapter.module.ts
@@ -1,14 +1,0 @@
-import { CREDENTIAL_CONFIG_YML_ADAPTER } from '@common/enums/providers';
-import { Module } from '@nestjs/common';
-import { CredentialConfigYmlAdapter } from './credential.config.yml.adapter';
-
-@Module({
-  providers: [
-    {
-      provide: CREDENTIAL_CONFIG_YML_ADAPTER,
-      useClass: CredentialConfigYmlAdapter,
-    },
-  ],
-  exports: [CREDENTIAL_CONFIG_YML_ADAPTER],
-})
-export class CredentialConfigAdapterModule {}
